refactor: (tobe) 메뉴 이름 Value Object 분리
