+++ src/main/java/kitchenpos/common/domain/Profanities.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.domain.tobe.domain;
+package kitchenpos.common.domain;
 
 public interface Profanities {
