+++ src/main/java/kitchenpos/menus/domain/tobe/domain/menu/DisplayedName.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.domain.tobe.domain;
+package kitchenpos.menus.domain.tobe.domain.menu;
 
 import java.util.Objects;
 import javax.persistence.Column;
@@ -21,16 +21,12 @@ public class DisplayedName {
 
     private void validate(final String value, final Profanities profanities) {
         if (Objects.isNull(value) || value.isEmpty()) {
-            throw new IllegalArgumentException("상품의 이름이 올바르지 않으면 등록할 수 없습니다.");
+            throw new IllegalArgumentException("메뉴의 이름이 올바르지 않으면 등록할 수 없습니다.");
         }
 
         if (profanities.contains(value)) {
-            throw new IllegalArgumentException("상품의 이름에는 비속어가 포함될 수 없습니다.");
+            throw new IllegalArgumentException("메뉴의 이름에는 비속어가 포함될 수 없습니다.");
         }
-    }
-
-    public String getDisplayedName() {
-        return this.value;
     }
 
     @Override
